Added volatile overloads for stk_simd
