+++ packages/stk/stk_simd/stk_simd/disimd/DISimdDouble.hpp
@@ -56,6 +56,12 @@ struct Double {
     : _data(x._data) {
   }
 
+#ifdef STK_VOLATILE_SIMD
+  STK_MATH_FORCE_INLINE Double(const volatile Double& x)
+    : _data(x._data) {
+  }
+#endif
+
   STK_MATH_FORCE_INLINE Double& operator= (const Double& x) {
     _data = x._data;
     return *this;
@@ -71,6 +77,12 @@ struct Double {
     _data += a._data;
     return *this;
   }
+
+#ifdef STK_VOLATILE_SIMD
+  STK_MATH_FORCE_INLINE void operator+= (const volatile Double& a) volatile {
+    _data.plus_equals(a._data);
+  }
+#endif
 
   STK_MATH_FORCE_INLINE Double& operator-= (const Double& a) {
     _data -= a._data;
